Provenance
New export tab “Provenance” in HTML uses new function outputProv() to
generate W3C prov-n notation, and the “Generate Graph” button calls the
new cgi script prov which uses the provconvert software to generate an
SVG graph.

Also, the “Dice” algorithm button calls “doDiceWalk()” to do a random
walk, as a variation on random, to help illustrate Markov approaches.
There is also a doRandomWalk() function, created during development of
doDiceWalk(), but this is unused at present.

Finally, the csvmidi CGI script now returns the midi file as
attachment, instead of a page containing a link. The old code remains
in the script for reuse.